Menu completo a prova de Bertinho

+++ leitor_configs.c
@@ -55,7 +55,7 @@ LISTA_COMBOIOS * preenche_comboio(char *aux_string, int *aux_int, LISTA_COMBOIOS
   novo_boio->boio.dim = aux_int[0];
   novo_boio->boio.origem = procura_ponto(aux_string+20, aux_string+30, topo_lista_linhas);
   novo_boio->boio.tempo_spawn = aux_int[1];
-  novo_boio->boio.veloc = aux_int[2] / FPS;
+  novo_boio->boio.veloc = (float) aux_int[2] / FPS;
 
   novo_boio->pr=topo_lista_comboios;
   topo_lista_comboios=novo_boio;
